Created memory pool class for handling buffers
